updated config example build tool to use consolidated list of files

+++ _doc_tools/build_examples.py
@@ -2,18 +2,15 @@ import filecmp
 import os
 import shutil
 
-source_paths = ['../../mpf/mpf/tests/machine_files',
-                '../../mpf-mc/mpfmc/tests/machine_files']
+source_config_paths = ['../../mpf/mpf/tests/machine_files',
+                       '../../mpf-mc/mpfmc/tests/machine_files']
 
-existing_paths = ['../example_configs/mpf_configs',
-                  '../example_configs/mpfmc_configs']
-
-file_types = ['.yaml', '.py']
-
-paths_to_ignore = ['data']
-
+local_config_path = '../example_configs'
 examples_root = '../examples'
 examples_index = '../examples/index.rst'
+
+file_types = ['.yaml', '.py']
+paths_to_ignore = ['data']
 
 
 class ExampleBuilder(object):
@@ -43,7 +40,7 @@ class ExampleBuilder(object):
         self.write_index()
 
     def create_source_file_list(self):
-        for source_path in source_paths:
+        for source_path in source_config_paths:
             for path, _, files in os.walk(source_path):
                 for file in files:
 
@@ -60,20 +57,21 @@ class ExampleBuilder(object):
                     self.source_files.append((full_path, partial_path))
 
     def create_existing_file_list(self):
-        for existing_path in existing_paths:
-            for path, _, files in os.walk(existing_path):
-                for file in files:
+        for path, _, files in os.walk(local_config_path):
+            for file in files:
 
-                    if os.path.splitext(file)[1] not in file_types:
-                        continue
+                if os.path.splitext(file)[1] not in file_types:
+                    continue
 
-                    full_path = os.path.join(path, file)
+                full_path = os.path.join(path, file)
 
-                    if self._check_ignored(full_path):
-                        continue
+                if self._check_ignored(full_path):
+                    continue
 
-                    partial_path = full_path.replace(existing_path, '')
-                    self.existing_files.append((full_path, partial_path))
+                partial_path = full_path.replace(local_config_path, '')
+                self.existing_files.append((full_path, partial_path))
+
+        # print(self.existing_files)
 
     def _check_ignored(self, path):
         for ignore_path in paths_to_ignore:
@@ -87,25 +85,16 @@ class ExampleBuilder(object):
             if file in [x[1] for x in self.existing_files]:
                 self.files_in_both.append((full_path, file))
 
-                for index, path in enumerate(source_paths):
-                    if full_path.startswith(path):
+                if not filecmp.cmp(full_path, local_config_path + file):
 
-                        if not filecmp.cmp(
-                                source_paths[index] + file,
-                                existing_paths[index] + file):
-
-                            self.changed_files.append(
-                                (source_paths[index] + file,
-                                 os.path.split(existing_paths[index] + file)[0]))
+                    self.changed_files.append(
+                        (full_path,
+                         os.path.split(local_config_path + file)[0]))
 
             else:
-
-                for index, path in enumerate(source_paths):
-                    if full_path.startswith(path):
-
-                        self.new_files.append(
-                            (source_paths[index] + file,
-                             os.path.split(existing_paths[index] + file)[0]))
+                self.new_files.append(
+                    (full_path,
+                     os.path.split(local_config_path + file)[0]))
 
         for full_path, file in self.existing_files:
             if file not in [x[1] for x in self.source_files]:
@@ -116,60 +105,69 @@ class ExampleBuilder(object):
         # copy new files
         for source, target in self.new_files:
             os.makedirs(target, exist_ok=True)
-            print("+++++", source, target)
+            target_file_with_path = os.path.join(target, os.path.split(source)[1])
+
+            # check if the target file exists:
+            if os.path.isfile(target_file_with_path):
+                print("WARNING: Duplicate file found:", target_file_with_path)
+                continue
+
+            print("<NEW> {} -> {}".format(source, target_file_with_path))
             shutil.copy(source, target)
 
         # copy changed files
         for source, target in self.changed_files:
             os.makedirs(target, exist_ok=True)
-            print(">>>>>", source, target)
+            target_file_with_path = os.path.join(target, os.path.split(source)[1])
+
+            print("<UPDATED> {} -> {}".format(source, target_file_with_path))
             shutil.copy(source, target)
 
         # delete removed files
         for file, _ in self.removed_files:
-            print("-----", file)
+            print("<REMOVED>", file)
             os.remove(file)
 
         if self.removed_files:
             self.remove_empty_dirs()
 
     def remove_empty_dirs(self):
-        for path in existing_paths:
-            empty_count = 0
-            used_count = 0
+        empty_count = 0
+        used_count = 0
 
-            for curdir, subdirs, files in os.walk(path):
-                if len(subdirs) == 0 and len(files) == 0:
-                    empty_count += 1
-                    os.rmdir(curdir)
-                elif len(subdirs) > 0 and len(files) > 0:
-                    used_count += 1
+        for curdir, subdirs, files in os.walk(local_config_path):
+            if len(subdirs) == 0 and len(files) == 0:
+                empty_count += 1
+                os.rmdir(curdir)
+            elif len(subdirs) > 0 and len(files) > 0:
+                used_count += 1
 
     def empty_existing_examples_folder(self):
         shutil.rmtree(examples_root)
         os.makedirs(examples_root)
 
     def create_examples_sections(self):
-        for path in existing_paths:
-            for dir in os.listdir(path):
-                self.examples_sections[dir] = dict()
-                self.examples_sections[dir]['config'] = list()
-                self.examples_sections[dir]['modes'] = list()
-                self.examples_sections[dir]['shows'] = list()
+        for dir in os.listdir(local_config_path):
+            self.examples_sections[dir] = dict()
+            self.examples_sections[dir]['config'] = list()
+            self.examples_sections[dir]['modes'] = list()
+            self.examples_sections[dir]['shows'] = list()
 
-            for full_path, dirs, files in os.walk(path):
-                rel_path = full_path.replace(path, '')
+        for full_path, dirs, files in os.walk(local_config_path):
+            rel_path = full_path.replace(local_config_path, '')
 
-                if files:
-                    folders = rel_path.lstrip('/').split('/')
-                    section = folders[0]
-                    folder = folders[1]
+            print (full_path, rel_path, dirs, files)
 
-                    if (section in self.examples_sections and
-                            folder in self.examples_sections[section]):
-                        for file in files:
-                            self.examples_sections[section][folder] = (
-                                os.path.join(rel_path, file))
+            if files:
+                folders = rel_path.lstrip('/').split('/')
+                section = folders[0]
+                folder = folders[1]
+
+                if (section in self.examples_sections and
+                        folder in self.examples_sections[section]):
+                    for file in files:
+                        self.examples_sections[section][folder] = (
+                            os.path.join(rel_path, file))
 
         import pprint
         pprint.pprint(self.examples_sections)
